Broadcasts on hover for feed page
Still need work on getting it to work for favorites. Its not working because dayview (which is called from feed) calls getGamesForDate() which links to the /scoreboard endpoint containing the broadcast data, while in favorites getTeamGames() is called which doesnt have that data so that would need to be fixed for it to show properly on favorites.